# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Network Clone",
        "description": "Social Network Clone API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "feed/friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get user friends",
                "description": "Returns user friends",
                "operationId": "getUserFriends",
                "responses": {
                    "200": {
                        "description": "Got user friends",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/friends/{userID}": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get user friends by user ID",
                "description": "Returns user friends by user ID",
                "operationId": "getUserFriendsById",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Got user friends",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Friends"
                ],
                "summary": "Update friendship",
                "description": "Update friendship",
                "operationId": "updateFriendship",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "accepted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Friendship updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Send friend request",
                "description": "Send friend request",
                "operationId": "sendFriendRequest",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request sent",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Remove friend",
                "description": "Remove friend",
                "operationId": "removeFriend",
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend removed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "feed/friends/requests": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get friend requests",
                "description": "Get friend requests",
                "operationId": "getFriendRequests",
                "responses": {
                    "200": {
                        "description": "Got friend requests",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/friends/blocked": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get blocked friends",
                "description": "Get blocked friends",
                "operationId": "getBlockedFriends",
                "responses": {
                    "200": {
                        "description": "Got blocked friends",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/friends/declined": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get declined friends",
                "description": "Get declined friends",
                "operationId": "getDeclinedFriends",
                "responses": {
                    "200": {
                        "description": "Got declined friends",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/friends/pending": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get pending friends",
                "description": "Get pending friends",
                "operationId": "getPendingFriends",
                "responses": {
                    "200": {
                        "description": "Got pending friends",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
...
}